Little modification modify space check logic, now check every tick.
# ANNOUNCEMENT #
This is a release candidate version(abbreviated as RC), the RCs for 1.18.2 will be 3 versions, only one of them will become the final release version. Which exactly to be the final version depends on users' response, You can vote your favorite version at my page from bilibili.com.

Main feature to this RC-2 version
Only environment & item check are dealt on an extra thread. Planting is dealt on server thread.